requirements.txt
flask
gunicorn
python-redmine
pygithub
requests
